pc/robotic_arm_zyy/virtual3d_robotic_arm_zyy/requirements.txt
PyOpenGL
PyGame
pyserial==3.5b0
pynput
